Aircraft_carrier_tower-main/pyproject.toml
[project]
name = "aircraft-carrier-tower"
version = "1.0.0"
description = "航模地面站控制系统"
authors = [
    {name = "Developer", email = "developer@example.com"}
]
dependencies = [
    "pyserial>=3.5",
    "pynput>=1.7.6",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
aircraft-carrier-tower = {path = "."}
